Implement basic class hierarchy and verification (+ Change line-separators to LF)

+++ updatedata/organizer/cleanup.py
@@ -5,4 +5,5 @@ import shutil
 # Cleans up tmp folder
 def cleanup():
     tmp_path = os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "tmp"))
-    shutil.rmtree(tmp_path)
+    if os.path.isdir(tmp_path):
+        shutil.rmtree(tmp_path)
